♻️(refactor): cleanup directory
